Relocated the Context constant to the test harness
