+++ src/lambdas/FileCoordinator/test/index.test.ts
@@ -2,9 +2,8 @@ import * as sinon from 'sinon'
 import * as DynamoDB from '../../../lib/vendor/AWS/DynamoDB'
 import * as StepFunctions from '../../../lib/vendor/AWS/StepFunctions'
 import chai from 'chai'
-import {getFixture} from '../../../util/mocha-setup'
+import {getFixture, testContext} from '../../../util/mocha-setup'
 import {handler} from '../src'
-import {testContext} from '../../../util/constants'
 const expect = chai.expect
 const localFixture = getFixture.bind(null, __dirname)
